AI Chatbot/requirements.txt
# Core FastAPI Dependencies
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# Pydantic (compatible with Python 3.14)
pydantic>=2.10.0
pydantic-settings>=2.0.0

# LangChain Dependencies (latest compatible versions)
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-core>=0.3.0
langchain-community>=0.3.0

# OpenAI
openai>=1.50.0

# Google Gemini (Generative AI client)
google-generativeai>=0.6.0,<2.0.0

# Environment Management
python-dotenv>=1.0.0

# HTTP Requests
requests>=2.32.0
httpx>=0.27.0

# Additional Utilities
python-multipart>=0.0.9
aiofiles>=24.0.0

# Optional: Database Support (uncomment when needed)
# psycopg2-binary>=2.9.9  # PostgreSQL
# pymongo>=4.8.0           # MongoDB
# sqlalchemy>=2.0.35       # ORM

# Optional: Monitoring & Logging (uncomment when needed)
# prometheus-client>=0.21.0
# python-json-logger>=2.0.7
